feat: Separacion de Logica y Visual dentro de los archivos
