+++ nonogram/__init__.py
@@ -1,2 +1,4 @@
 from .image_to_matrix import ImageToMatrix
 from .excepciones_imagenes import NoExisteMatrizError, ImagenError
+from .calculo_puntaje import calculador_puntaje
+from .scripts_de_prueba.pruebaspygame import get_num_arriba, get_num_lado
