+++ nonogram/visual/__init__.py
@@ -1,0 +1,3 @@
+from .tablero_visual import TableroVisual
+from .boton import Boton
+from .colores import Colores
