+++ nonogram/visual/colores.py
